Fix integer division for Python 3. Fixes partially PIPE-8611
git-svn-id: http://svnhq2.hq.eso.org/p2/branches/Pipelines/gravity/gravity_dev/gravityp@247884 318c8218-2dab-440d-9263-6f70bac63e5b

+++ reflex/gravity_plot_common.py
@@ -30,7 +30,7 @@ class GravityInstrumentConf:
     this_lab_input = self.getLabInputIndex(gravity_input)
     ncol_ot = self.conf.getTableNcols('OPTICAL_TRAIN')
     tel_name_col = self.conf.readTableColumn('OPTICAL_TRAIN', 'TEL_NAME')
-    for col in range(1, (ncol_ot - 2 ) / 2 + 1) :
+    for col in range(1, int((ncol_ot - 2 ) / 2) + 1) :
       colname = "OPTI_NAME{0}".format(col)
       opti_elem = self.conf.readTableColumn('OPTICAL_TRAIN', colname)
       if(opti_elem[0] == 'LAB_INPUT1'):
